tests(*) refactor mock_modules into spec.fixtures.mocker
Refactor the `mock_modules` functionality from the kong.runloop.handler
test file into a utility module to be reused by other unit test modules.
